requirements.txt
six
unicodecsv
pyyaml
datapackage